Ship yabridgectl and fix file locations
yabridgectl was part of the source code but wasn't being compiled. It's quite convenient to configure yabridge
yabridgectl expects vst2 and vst3 wrappers in the same folder. moved libyabridge-vst3.so to the /vst folder (instead of /vst3)
yabridge-{host,group}.exe expects yabridge-{host,group}.exe.so in the same folder. Moved from lib64 to bin